+++ src/serde.rs
@@ -2,8 +2,7 @@
 
 use crate::{
     core::DbError,
-    db_row::{DbRow, JsonRow},
-    db_value::{DbValue, JsonValue},
+    db_value::{DbRow, DbValue, JsonRow, JsonValue},
 };
 use serde::{
     Deserialize, Serialize,
