+++ src/tokio_postgres.rs
@@ -5,8 +5,7 @@ use crate::{
     cache::{CachingStrategy, clear_cache_for_affected_tables, clear_cache_for_dropped_tables},
     core::{DbError, DbQuery},
     db_kind::{DbKind, MAX_PARAMS_POSTGRES, PostgreSQLKind},
-    db_row::{DbRow, DbRows, IntoDbRows},
-    db_value::{DbParams, DbValue, IntoDbParams, JsonValue},
+    db_value::{DbParams, DbRow, DbRows, DbValue, IntoDbParams, IntoDbRows, JsonValue},
     parse::validate_table_name,
     shared::{EditType, edit},
 };
