+++ tests/axum.rs
@@ -4,7 +4,7 @@ use axum::{
     response::{Html, IntoResponse},
     routing::get,
 };
-use rltbl_db::{any::AnyPool, core::DbQuery, db_row, db_row::DbRow};
+use rltbl_db::{any::AnyPool, core::DbQuery, db_row, db_value::DbRow};
 use std::{marker::Sync, sync::Arc};
 use tower_service::Service;
